refactor(userspace/libsinsp): polish name of filtercheck sources
