+++ userspace/libsinsp/filtercheck_container.h
