+++ userspace/libsinsp/filtercheck_fspath.cpp
@@ -17,10 +17,12 @@ limitations under the License.
 */
 
 #include <filesystem>
+#include "filtercheck_fspath.h"
+#include "filtercheck_event.h"
+#include "filtercheck_fd.h"
 #include "sinsp.h"
 #include "sinsp_int.h"
 #include "filter.h"
-#include "filterchecks.h"
 
 using namespace std;
