+++ userspace/libsinsp/filtercheck_fspath.h
