+++ userspace/libsinsp/filtercheck_mesos.cpp
@@ -16,11 +16,10 @@ limitations under the License.
 
 */
 
-
+#include "filtercheck_mesos.h"
 #include "sinsp.h"
 #include "sinsp_int.h"
 #include "filter.h"
-#include "filterchecks.h"
 
 using namespace std;
