+++ userspace/libsinsp/filtercheck_rawstring.h
