+++ userspace/libsinsp/filtercheck_syslog.cpp
@@ -16,10 +16,10 @@ limitations under the License.
 
 */
 
+#include "filtercheck_syslog.h"
 #include "sinsp.h"
 #include "sinsp_int.h"
 #include "filter.h"
-#include "filterchecks.h"
 #include "protodecoder.h"
 
 using namespace std;
