+++ userspace/libsinsp/filtercheck_syslog.h
