+++ userspace/libsinsp/filtercheck_thread.h
