+++ userspace/libsinsp/filtercheck_tracer.cpp
@@ -18,10 +18,11 @@ limitations under the License.
 
 #include <math.h>
 
+#include "filtercheck_tracer.h"
+#include "filtercheck_reference.h"
 #include "sinsp.h"
 #include "sinsp_int.h"
 #include "filter.h"
-#include "filterchecks.h"
 #include "tracers.h"
 #include "strl.h"
