+++ userspace/libsinsp/filtercheck_utils.cpp
@@ -16,10 +16,10 @@ limitations under the License.
 
 */
 
+#include "filtercheck_utils.h"
 #include "sinsp.h"
 #include "sinsp_int.h"
 #include "filter.h"
-#include "filterchecks.h"
 
 using namespace std;
