+++ userspace/libsinsp/filtercheck_utils.h
