+++ userspace/libsinsp/filterchecks.h
@@ -22,19 +22,20 @@ limitations under the License.
 // reasons, as all filterchecks used to be defined in this file
 
 #include "sinsp_filtercheck.h"
-#include "filter/checks/event.h"
-#include "filter/checks/evtin.h"
-#include "filter/checks/fd.h"
-#include "filter/checks/fdlist.h"
-#include "filter/checks/fspath.h"
-#include "filter/checks/gen_event.h"
-#include "filter/checks/group.h"
-#include "filter/checks/k8s.h"
-#include "filter/checks/mesos.h"
-#include "filter/checks/rawstring.h"
-#include "filter/checks/reference.h"
-#include "filter/checks/syslog.h"
-#include "filter/checks/thread.h"
-#include "filter/checks/tracer.h"
-#include "filter/checks/user.h"
-#include "filter/checks/utils.h"
+#include "filtercheck_container.h"
+#include "filtercheck_event.h"
+#include "filtercheck_evtin.h"
+#include "filtercheck_fd.h"
+#include "filtercheck_fdlist.h"
+#include "filtercheck_fspath.h"
+#include "filtercheck_gen_event.h"
+#include "filtercheck_group.h"
+#include "filtercheck_k8s.h"
+#include "filtercheck_mesos.h"
+#include "filtercheck_rawstring.h"
+#include "filtercheck_reference.h"
+#include "filtercheck_syslog.h"
+#include "filtercheck_thread.h"
+#include "filtercheck_tracer.h"
+#include "filtercheck_user.h"
+#include "filtercheck_utils.h"
